Linear.Logic

Description

Linear Logic for Constructive Mathematics by Michael Shulman provides a principled take on this topic. There he constructs an embedding of an affine logic into an intuitionistic logic via a Chu construction.

However, that version of things was only able to express an 'affine logic' where the pairs \(\top\) and (), \(\bot\) and Void are made to coincide.

Reconstructing this technique on top of Linear Haskell allows us to construct a full intuitionistic linear logic, while retaining Shulman's style of refutation.

Additive	Multiplicative

Conjunction	(&) w/ unit Top	(,) w/ unit ()

Disjunction	Either w/ unit Void	(⅋) w/ unit Bot

Either (or (+)) takes the place of the traditional \(\oplus\)

(,) (or (*)) takes the place of the traditional \(\otimes\)

To use the alias for (*), make sure to enable {-# LANGUAGE NoStarIsType #-}

Negative polarity connectives are RuntimeRep polymorphic, but only currently have Prop instances defined for 'LiftedRep

Documentation

class (Prop (Not a), Not (Not a) ~ a) => Prop (a :: TYPE k) where

Associated Types

type Not (a :: TYPE k) = (c :: TYPE k) | c -> a

\(a^\bot\). The type of refutations of \(a\)

\(a^{\bot^\bot} \) = \(a\)

Methods

(!=) :: a -> Not a -> r

\(a\) and \(a^\bot\) together yield a contradiction.

(!=) :: a %1 -> Not a %1 -> r

newtype a & b infixr 3

Constructors

With (forall k (c :: TYPE k). Y a b c -> c)

data Top where

Top can hold any unwanted environment, which allows it to work as a unit for (&).

Constructors

Top :: a -> Top
Top :: a %1 -> Top

type With = (&)

with :: forall i j (a :: TYPE i) (b :: TYPE j). (forall k (c :: TYPE k). Y a b c -> c) -> a & b

Introduce a With/(&) connective.

Usage:

with case
  L -> ...
  R -> ...

with :: (forall c. Y a b c -> c) %1 -> a & b

withL :: (a & b) -> a

Eliminate a With/(&) connective and extract the left choice.

withL (with case L -> x; R -> y) ≡ x

withL :: a & b %1 -> a

withR :: (a & b) -> b

Eliminate a With/(&) connective and extract the right choice.

withR (with case L -> x; R -> y) ≡ y

withR :: a & b %1 -> b

type (+) = Either infixr 2

data Void

Uninhabited data type

Since: base-4.8.0.0

data Either a b

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

Expand

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Constructors

Left a

Right b

type * = (,) infixr 3

newtype (a :: TYPE i) ⅋ (b :: TYPE j) infixr 2

\(\par\) is multiplicative disjunction.

Constructors

Par (forall k (c :: TYPE k). Y (Not b -> a) (Not a -> b) c -> c)

data Bot where

Bot acts as a unit for (⅋)

Constructors

Bot :: (forall a. a) -> Bot

type Par = (⅋)

par :: forall i j (a :: TYPE i) (b :: TYPE j). (forall k (c :: TYPE k). Y (Not b -> a) (Not a -> b) c -> c) -> a ⅋ b

Introduce a par/(⅋) connective.

Usage:

par case
  L -> ...
  R -> ...

When developing using holes, you may want to temporarily substitute unsafePar until all the holes have been solved, then putting this in instead once everything typechecks.

par :: (forall c. Y (Not b %1 -> a) (Not a %1 -> b) c %1 -> c) %1 -> a ⅋ b

parL :: (a ⅋ b) -> Not b -> a

Eliminate a par/(⅋) connective, given refutation of the b, supply proof of a.

parL (par case L -> x; R -> y) ≡ x

parL :: a ⅋ b %1 -> Not b %1 -> a

parR :: (a ⅋ b) -> Not a -> b

Eliminate a par/(⅋) connective, given refutation of the a, supply proof of b.

parR (par case L -> x; R -> y) ≡ y

parR :: a ⅋ b %1 -> Not a %1 -> b

type (⊸) p q = Not p ⅋ q infixr 0

\(\multimap\) is defined in terms of \(⅋\)

fun :: forall a b. Prop a => (a ⊸ b) -> a -> b

Derive a linear function from a linear logic impliciation.

fun :: forall a b. Prop a => (a ⊸ b) %1 -> a %1 -> b

unfun :: forall a b. (a ⊸ b) -> Not b -> Not a

Derive a linear function for the contrapositive from a linear logic impliciation.

unfun :: forall a b. Prop a => (a ⊸ b) %1 -> a %1 -> b

data Nofun a b

The refutation of a linear haskell arrow.

Constructors

Nofun a (Not b)

data Ur a where

Constructors

Ur :: forall a. a -> Ur a

newtype No a

Heyting negation

Constructors

No (forall r. a -> r)

extractUr :: forall p. Prop p => Ur p ⊸ p

duplicateUr :: forall p. Ur p ⊸ Ur (Ur p)

weakenUr :: forall p q. Prop p => p ⊸ (Ur q ⊸ p)

weakenUr :: forall p q. Prop p => p ⊸ (Ur q ⊸ p)

distUr :: forall p q. Prop p => Ur (p ⊸ q) ⊸ (Ur p ⊸ Ur q)

contractUr :: (Prop p, Prop q) => (Ur p ⊸ (Ur p ⊸ q)) ⊸ (Ur p ⊸ q)

newtype WhyNot a

The \(?a\) or "why not?" modality.

Constructors

WhyNot (forall r. Not a -> r)

because :: WhyNot a -> Not a -> r

whyNot :: (forall r. Not a -> r) -> WhyNot a

newtype Why a

The refutation of \(?a\)

Constructors

Why (Not a)

why :: Not a -> Why a

runWhy :: Why a -> Not a

returnWhyNot :: forall p. Prop p => p ⊸ WhyNot p

joinWhyNot :: forall p. Prop p => WhyNot (WhyNot p) ⊸ WhyNot p

data Y a b c where

Bundled Patterns

pattern R :: forall i j k (a :: i) (b :: j) (c :: k). () => b ~~ c => Y a b c

pattern L :: forall i j k (a :: i) (b :: j) (c :: k). () => a ~~ c => Y a b c