Linear.Logic.Unsafe

Documentation

toLinear :: forall (r1 :: RuntimeRep) (r2 :: RuntimeRep) (a :: TYPE r1) (b :: TYPE r2) (p :: Multiplicity). (a -> b) -> a -> b

unsafePar :: (forall c. Y (Not b -> a) (Not a -> b) c -> c) -> a ⅋ b

When developing proofs it is often handy to lie about multiplicity, while you are still filling in holes. This enables that workflow. When you are done, you can convert from unsafePar back to par

toLinear can also come in handy temporarily during this process.